fix: allow using settings.DJELME_BACKENDS before app ready
