+++ elasticsearch_metrics/apps.py
@@ -2,12 +2,10 @@ import collections
 
 from django.apps import AppConfig
 from django.conf import settings
-from django.core.exceptions import ImproperlyConfigured
 from django.utils.module_loading import autodiscover_modules
 
 from elasticsearch_metrics.registry import djelme_registry
 
-BACKENDS_SETTING = "DJELME_BACKENDS"
 AUTOSETUP_SETTING = "DJELME_AUTOSETUP"
 
 
@@ -17,14 +15,15 @@ class ElasticsearchMetricsConfig(AppConfig):
     def ready(self) -> None:
         # load backends settings
         _backend_names_by_module = collections.defaultdict(list)
-        for _backend_name, _imp_module_name, _imp_kwargs in _each_backend_setting():
+        for (
+            _backend_name,
+            _imp_module_name,
+            _,
+        ) in djelme_registry.each_backend_settings():
             _backend_names_by_module[_imp_module_name].append(_backend_name)
-            djelme_registry.register_backend(
-                _backend_name, _imp_module_name, _imp_kwargs
-            )
         # discover any `foo.metrics` in installed apps
         autodiscover_modules("metrics")
-        # call `djelme_when_ready` on each imp module
+        # call `djelme_when_ready` for each imp module (only once)
         for _imp_module_name, _backend_names in _backend_names_by_module.items():
             _imp_module = djelme_registry.get_imp_module(_imp_module_name)
             _imp_module.djelme_when_ready(
@@ -39,28 +38,3 @@ class ElasticsearchMetricsConfig(AppConfig):
                 _recordtypes,
             ) in djelme_registry.each_recordtype_by_backend():
                 djelme_registry.get_backend(_backend_name).djelme_setup(_recordtypes)
-
-
-###
-# accessing django settings
-
-
-def _each_backend_setting() -> (
-    collections.abc.Iterator[tuple[str, str, dict[str, str]]]
-):
-    for _backend_name, _backend_imps in getattr(settings, BACKENDS_SETTING, {}).items():
-        if len(_backend_imps) > 1:
-            raise ImproperlyConfigured(
-                BACKENDS_SETTING, "only one imp per backend, at the moment"
-            )
-        for _imp_module_name, _imp_kwargs in _backend_imps.items():
-            if not (
-                isinstance(_imp_module_name, str)
-                and isinstance(_imp_kwargs, dict)
-                and all(isinstance(_k, str) for _k in _imp_kwargs.keys())
-            ):
-                raise ImproperlyConfigured(
-                    BACKENDS_SETTING,
-                    'expected {"mybackend": {"imp.path": {"imp_config": "..."}}}',
-                )
-            yield (_backend_name, _imp_module_name, _imp_kwargs)
